refactor: rename SymAiContext child element classes with prefix
- Rename Stream to SymAiContextStream
- Rename Message to SymAiContextMessage
- Rename Attachment to SymAiContextAttachment
- Add optional from and to timestamp attributes to SymAiContextStream
- Add comprehensive Javadoc documentation to all SymAiContext-related classes
