+++ src/main/java/org/finos/symphony/messageml/messagemlutils/MessageMLParser.java
@@ -18,7 +18,7 @@ import org.apache.commons.lang3.tuple.Pair;
 import org.finos.symphony.messageml.messagemlutils.bi.BiContext;
 import org.finos.symphony.messageml.messagemlutils.bi.BiFields;
 import org.finos.symphony.messageml.messagemlutils.bi.BiItem;
-import org.finos.symphony.messageml.messagemlutils.elements.Attachment;
+import org.finos.symphony.messageml.messagemlutils.elements.SymAiContextAttachment;
 import org.finos.symphony.messageml.messagemlutils.elements.Bold;
 import org.finos.symphony.messageml.messagemlutils.elements.BulletList;
 import org.finos.symphony.messageml.messagemlutils.elements.Button;
@@ -69,8 +69,8 @@ import org.finos.symphony.messageml.messagemlutils.elements.SplittableElement;
 import org.finos.symphony.messageml.messagemlutils.elements.Subscript;
 import org.finos.symphony.messageml.messagemlutils.elements.Superscript;
 import org.finos.symphony.messageml.messagemlutils.elements.SymAiContext;
-import org.finos.symphony.messageml.messagemlutils.elements.Stream;
-import org.finos.symphony.messageml.messagemlutils.elements.Message;
+import org.finos.symphony.messageml.messagemlutils.elements.SymAiContextStream;
+import org.finos.symphony.messageml.messagemlutils.elements.SymAiContextMessage;
 import org.finos.symphony.messageml.messagemlutils.elements.Table;
 import org.finos.symphony.messageml.messagemlutils.elements.TableBody;
 import org.finos.symphony.messageml.messagemlutils.elements.TableCell;
@@ -498,14 +498,14 @@ public class MessageMLParser {
       case SymAiContext.MESSAGEML_TAG:
         return new SymAiContext(parent, ++index);
 
-      case Stream.MESSAGEML_TAG:
-        return new Stream(parent);
+      case SymAiContextStream.MESSAGEML_TAG:
+        return new SymAiContextStream(parent);
 
-      case Message.MESSAGEML_TAG:
-        return new Message(parent);
+      case SymAiContextMessage.MESSAGEML_TAG:
+        return new SymAiContextMessage(parent);
 
-      case Attachment.MESSAGEML_TAG:
-        return new Attachment(parent);
+      case SymAiContextAttachment.MESSAGEML_TAG:
+        return new SymAiContextAttachment(parent);
 
       case Bold.MESSAGEML_TAG:
         return new Bold(parent);
